organized app.core folder
