+++ app/core/__init__.py
@@ -1,5 +1,5 @@
 """Core application modules."""
 
-from .logging import get_logger, setup_logging
+from .logging.logging import get_logger, setup_logging
 
 __all__ = ["setup_logging", "get_logger"]
