+++ app/repositories/base_repo.py
@@ -1,6 +1,6 @@
 """Base repository class."""
 
-from app.core.logging import get_logger
+from app.core.logging.logging import get_logger
 
 logger = get_logger(__name__)
