+++ app/repositories/chat_repo.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 from sqlalchemy.orm import Session
 
-from app.core.logging import get_logger
+from app.core.logging.logging import get_logger
 from app.models.chat import Chat, DirectMessage, GroupChat
 from app.models.membership import MemberRole, Membership
 from app.repositories.base_repo import BaseRepo
